[MOD/#51] UserInfoResponse 수정
- kakaoId 가 Int 범위보다 커서 Long으로 수정

+++ app/src/main/java/umc/link/zip/data/dto/mypage/response/UserInfoResponse.kt
@@ -5,7 +5,7 @@ import umc.link.zip.domain.model.mypage.UserInfoModel
 data class UserInfoResponse(
     var userId: Int,
     var nickname: String,
-    var kakaoId: Int,
+    var kakaoId: Long,
     var createdAt: String
 ){
     fun toModel() = UserInfoModel (
